Move scheduling core to bootstrap module
Move all Netty-independent scheduling classes from core to bootstrap
under io.netty.loom.scheduler package. Eliminates the per-classloader
singleton problem in app servers (JBoss/Liberty) where multiple
classloaders would create duplicate carrier groups.

What moved to bootstrap:
- NettyScheduler (entry point, was in io.netty.loom.spi)
- EventLoopScheduler (carrier loop, run queue, work stealing)
- EventLoopSchedulerGroup (carrier pool)
- SchedulerJfrUtil + 4 JFR events
- jctools as a regular dependency (shade blocked by ASM/Java 27)

What changed:
- SPI eliminated: NettyScheduler directly routes onStart/onContinue
- Eager carrier pool creation in NettyScheduler constructor
- NettyScheduler reference passed explicitly (no partial publication)
- Tests moved to io.netty.loom.scheduler package for visibility
- SchedulerJfrUtil split: scheduler events in bootstrap, Netty I/O
  events (NettyJfrUtil) stay in core

JVM flag changed:
  -Djdk.virtualThreadScheduler.implClass=io.netty.loom.scheduler.NettyScheduler
