+++ bootstrap/src/main/java/io/netty/loom/scheduler/jfr/WorkStealEvent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 The Netty VirtualThread Scheduler Project
+ * Copyright 2026 The Netty VirtualThread Scheduler Project
  *
  * The Netty VirtualThread Scheduler Project licenses this file to you under the Apache License,
  * version 2.0 (the "License"); you may not use this file except in compliance with the
@@ -12,7 +12,7 @@
  * either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package io.netty.loom.jfr;
+package io.netty.loom.scheduler.jfr;
 
 import jdk.jfr.Category;
 import jdk.jfr.Description;
@@ -23,27 +23,31 @@ import jdk.jfr.Label;
 import jdk.jfr.Name;
 import jdk.jfr.StackTrace;
 
-@Name("io.netty.loom.VirtualThreadTaskRun")
-@Label("Virtual Thread Task Run")
-@Description("Virtual thread task execution by the scheduler.")
+@Name("io.netty.loom.WorkSteal")
+@Label("Work Steal")
+@Description("A virtual thread task was stolen from a busy carrier by an idle one.")
 @Category({"Netty", "Scheduler"})
 @Enabled(false)
 @StackTrace(false)
-public final class VirtualThreadTaskRunEvent extends Event {
+public final class WorkStealEvent extends Event {
 
-	private static final EventType EVENT_TYPE = EventType.getEventType(VirtualThreadTaskRunEvent.class);
+	private static final EventType EVENT_TYPE = EventType.getEventType(WorkStealEvent.class);
 
-	@Label("Carrier Thread")
-	public Thread carrierThread;
-
-	@Label("Virtual Thread")
+	@Label("Stolen Virtual Thread")
 	public Thread virtualThread;
 
-	@Label("Is Poller")
-	public boolean isPoller;
+	@Label("Source Carrier")
+	public Thread sourceCarrier;
 
-	@Label("Is Event Loop")
-	public boolean isEventLoop;
+	@Label("Stealer Carrier")
+	public Thread stealerCarrier;
+
+	@Label("Source Queue Depth")
+	public int sourceQueueDepth;
+
+	@Label("From Carrier Loop")
+	@Description("True if stolen from the carrier loop; false if stolen from the pinned poller via maybeYield.")
+	public boolean fromCarrierLoop;
 
 	public static boolean isEventEnabled() {
 		return EVENT_TYPE.isEnabled();
