+++ core/src/main/java/io/netty/loom/VirtualIoNioPollerEventLoopGroup.java
@@ -25,7 +25,8 @@ import io.netty.channel.IoEventLoop;
 import io.netty.channel.IoHandlerFactory;
 import io.netty.channel.ManualIoEventLoop;
 import io.netty.channel.MultiThreadIoEventLoopGroup;
-import io.netty.loom.spi.NettyScheduler;
+import io.netty.loom.scheduler.EventLoopScheduler;
+import io.netty.loom.scheduler.EventLoopSchedulerGroup;
 import io.netty.util.concurrent.FastThreadLocalThread;
 
 /**
@@ -71,19 +72,6 @@ public class VirtualIoNioPollerEventLoopGroup extends MultiThreadIoEventLoopGrou
 					"need " + nThreads + " free schedulers but not enough available in group of " + group.size());
 		}
 		return schedulers;
-	}
-
-	private static void validateNettyAvailability() {
-		if (!NettyScheduler.isAvailable()) {
-			if (!NettyScheduler.isInstalled()) {
-				throw new IllegalStateException(
-						"-Djdk.virtualThreadScheduler.implClass=io.netty.loom.spi.NettyScheduler is required");
-			}
-			throw new IllegalStateException(
-					"NettyScheduler bootstrap is installed but no NettySchedulerSpi provider was found. "
-							+ "Ensure the core module JAR is visible to the thread-context class loader "
-							+ "and META-INF/services/io.netty.loom.spi.NettySchedulerSpi is present.");
-		}
 	}
 
 	public java.util.concurrent.ThreadFactory vThreadFactoryOf(IoEventLoop eventLoop) {
@@ -145,7 +133,7 @@ public class VirtualIoNioPollerEventLoopGroup extends MultiThreadIoEventLoopGrou
 
 	private static int runIO(EventLoopScheduler scheduler, ManualIoEventLoop ioEventLoop, boolean canBlock,
 			AtomicBoolean pollerRunning) {
-		var event = SchedulerJfrUtil.beginRunIoEvent();
+		var event = NettyJfrUtil.beginRunIoEvent();
 		int ioEventsHandled;
 		boolean ranBlocking = false;
 		if (canBlock) {
@@ -160,21 +148,21 @@ public class VirtualIoNioPollerEventLoopGroup extends MultiThreadIoEventLoopGrou
 			ioEventsHandled = ioEventLoop.runNow(EventLoopScheduler.YIELD_DURATION_NS);
 		}
 		if (event != null) {
-			SchedulerJfrUtil.commitRunIoEvent(event, scheduler.carrierThread(), ranBlocking, ioEventsHandled);
+			NettyJfrUtil.commitRunIoEvent(event, scheduler.carrierThread(), ranBlocking, ioEventsHandled);
 		}
 		return ioEventsHandled;
 	}
 
 	private static int runNonBlockingTasks(EventLoopScheduler scheduler, ManualIoEventLoop ioEventLoop,
 			long deadlineNs) {
-		var event = SchedulerJfrUtil.beginRunTasksEvent();
+		var event = NettyJfrUtil.beginRunTasksEvent();
 		if (event == null) {
 			return ioEventLoop.runNonBlockingTasks(deadlineNs);
 		}
 		int queueDepthBefore = scheduler.runnableCount();
 		int tasksHandled = ioEventLoop.runNonBlockingTasks(deadlineNs);
 		int queueDepthAfter = scheduler.runnableCount();
-		SchedulerJfrUtil.commitRunTasksEvent(event, scheduler.carrierThread(), tasksHandled, queueDepthBefore,
+		NettyJfrUtil.commitRunTasksEvent(event, scheduler.carrierThread(), tasksHandled, queueDepthBefore,
 				queueDepthAfter);
 		return tasksHandled;
 	}
@@ -183,7 +171,6 @@ public class VirtualIoNioPollerEventLoopGroup extends MultiThreadIoEventLoopGrou
 	@Override
 	protected IoEventLoop newChild(Executor executor, IoHandlerFactory ioHandlerFactory,
 			@SuppressWarnings("unused") Object... args) {
-		validateNettyAvailability();
 		var schedulers = (EventLoopScheduler[]) args[0];
 		if (eventSchedulerMappings == null) {
 			eventSchedulerMappings = new IdentityHashMap<>(schedulers.length);
