--- core/src/main/resources/META-INF/services/io.netty.loom.spi.NettySchedulerSpi
@@ -1,1 +1,0 @@
-io.netty.loom.NettySchedulerProviderImpl
