+++ pom.xml
@@ -14,7 +14,7 @@
         <module>core</module>
         <module>benchmarks</module>
         <module>benchmark-runner</module>
-        <module>example-echo</module>
+        <module>spring-example</module>
         <module>concurrency-tests</module>
     </modules>
 
@@ -123,7 +123,7 @@
                                 <goal>test</goal>
                             </goals>
                             <configuration>
-                                <argLine>--enable-preview -Djdk.virtualThreadScheduler.implClass=io.netty.loom.spi.NettyScheduler -Djdk.traceVirtualThreadLocals=false -Dio.netty.loom.schedulers=2 --add-opens=java.base/java.lang=ALL-UNNAMED</argLine>
+                                <argLine>--enable-preview -Djdk.virtualThreadScheduler.implClass=io.netty.loom.scheduler.NettyScheduler -Djdk.traceVirtualThreadLocals=false -Dio.netty.loom.schedulers=2 --add-opens=java.base/java.lang=ALL-UNNAMED</argLine>
                             </configuration>
                         </execution>
                         <execution>
@@ -133,7 +133,7 @@
                                 <goal>test</goal>
                             </goals>
                             <configuration>
-                                <argLine>--enable-preview -Djdk.virtualThreadScheduler.implClass=io.netty.loom.spi.NettyScheduler -Djdk.traceVirtualThreadLocals=false -Dio.netty.loom.schedulers=2 --add-opens=java.base/java.lang=ALL-UNNAMED -Djdk.pollerMode=3</argLine>
+                                <argLine>--enable-preview -Djdk.virtualThreadScheduler.implClass=io.netty.loom.scheduler.NettyScheduler -Djdk.traceVirtualThreadLocals=false -Dio.netty.loom.schedulers=2 --add-opens=java.base/java.lang=ALL-UNNAMED -Djdk.pollerMode=3</argLine>
                             </configuration>
                         </execution>
                         <execution>
@@ -143,7 +143,7 @@
                                 <goal>test</goal>
                             </goals>
                             <configuration>
-                                <argLine>--enable-preview -Djdk.virtualThreadScheduler.implClass=io.netty.loom.spi.NettyScheduler -Djdk.traceVirtualThreadLocals=false -Dio.netty.loom.schedulers=2 --add-opens=java.base/java.lang=ALL-UNNAMED -Dio.netty.loom.workstealing.enabled=true -Dio.netty.loom.workstealing.unresponsive.ms=5 -Djdk.trackAllThreads=true</argLine>
+                                <argLine>--enable-preview -Djdk.virtualThreadScheduler.implClass=io.netty.loom.scheduler.NettyScheduler -Djdk.traceVirtualThreadLocals=false -Dio.netty.loom.schedulers=2 --add-opens=java.base/java.lang=ALL-UNNAMED -Dio.netty.loom.workstealing.enabled=true -Dio.netty.loom.workstealing.unresponsive.ms=5 -Djdk.trackAllThreads=true</argLine>
                                 <forkedProcessTimeoutInSeconds>120</forkedProcessTimeoutInSeconds>
                                 <forkedProcessExitTimeoutInSeconds>30</forkedProcessExitTimeoutInSeconds>
                             </configuration>
